+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [2.2.5] - 2024-01-15
+## [2.3.0] - 2024-01-15
 
 ### Added